First acceptance test that creates Basic Edit.
===
Right now running the test 'suite' will launch a GUI. That's a design
that will probably have to change, but for now let's march on.

The goal is to find a way to test that PyRoom is creating a 'session' to
keep track of all open buffers so that on reload it reloads that
buffer.